# jar: reduce redundant allocations and copies (2x-3x speedup)